Fix declaration of mblen

+++ build.sh
@@ -47,14 +47,14 @@ declare -r optflags='-w -O2'
 declare -r linkflags='-Xlinker -s'
 
 declare -ra targets=(
-	# 'mips64el-unknown-linux-android'
-	# 'mipsel-unknown-linux-android'
+	'mips64el-unknown-linux-android'
+	'mipsel-unknown-linux-android'
 	'aarch64-unknown-linux-android'
-	# 'riscv64-unknown-linux-android'
-	# 'x86_64-unknown-linux-android'
-	# 'i686-unknown-linux-android'
-	# 'armv5-unknown-linux-androideabi'
-	# 'armv7-unknown-linux-androideabi'
+	'riscv64-unknown-linux-android'
+	'x86_64-unknown-linux-android'
+	'i686-unknown-linux-android'
+	'armv5-unknown-linux-androideabi'
+	'armv7-unknown-linux-androideabi'
 )
 
 declare -ra versions=(
@@ -785,13 +785,15 @@ for triplet in "${targets[@]}"; do
 		ar="${toolchain_directory}/bin/${triplet}-ar"
 	fi
 	
-	cp \
-		"${workdir}/submodules/obggcc/tools/gcc-wrapper/fstream."{c,h} \
-		"${workdir}/tools"
-	
-	[ -f libpino.a ] && unlink libpino.a
-	${cc} -c "${workdir}/tools/"*'.c'
-	${ar} rcs 'libpino.a' 'mmap64.o'
+	if (( base_version < 21 )); then
+		cp \
+			"${workdir}/submodules/obggcc/tools/gcc-wrapper/fstream."{c,h} \
+			"${workdir}/tools"
+		
+		[ -f libpino.a ] && unlink libpino.a
+		${cc} -c "${workdir}/tools/"*'.c'
+		${ar} rcs 'libpino.a' 'mmap64.o'
+	fi
 	
 	if [[ "$(basename "${PWD}")" = 'lib64' ]]; then
 		mv ./* '../lib' || true
